backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Azure SDKs
azure-search-documents==11.6.0
azure-storage-blob==12.22.0
azure-identity==1.19.0
openai==1.54.0

# LangChain
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.0

# Document processing
pypdf2==3.0.1
pymupdf==1.24.14
python-docx==1.1.2

# Data handling
pandas==2.2.2
pydantic==2.9.2
pydantic-settings==2.5.2

# PDF generation for document export
reportlab==4.2.2

# Additional utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0

# Token counting and text processing
tiktoken==0.8.0
